# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fiber Banking API
  version: 1.0.0
  description: |
    Dokumentasi OpenAPI untuk API perbankan sederhana berbasis Go Fiber.

    Semua pesan response menggunakan bahasa Indonesia dan format standar:
    - Sukses: `{ data: T, status: true, message: string }`
    - Error: `{ status: false, message: string }`
servers:
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/staff/dashboard:
    get:
      tags: [Staff]
      summary: Ringkasan dashboard staff (admin, staff)
      description: Mengambil ringkasan customer, saldo, total income/outcome 30 hari, dan data chart harian.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
      responses:
        "200":
          $ref: "#/components/responses/DashboardSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/staff/customers:
    get:
      tags: [Staff]
      summary: List customers (admin, staff)
      description: |
        Mengambil daftar customer beserta rekening dan saldo. Bisa diakses role `staff` atau `admin`.

        role = `admin`, `staff`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PerPageQuery"
        - $ref: "#/components/parameters/CustomerStatusQuery"
      responses:
        "200":
          $ref: "#/components/responses/CustomerListSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [Staff]
      summary: Buat customer (admin, staff)
      description: Staff atau admin membuat customer beserta rekening awal bersaldo 0.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
      requestBody:
        $ref: "#/components/requestBodies/CustomerBody"
      responses:
        "201":
          $ref: "#/components/responses/CustomerSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/staff/customers/options:
    get:
      tags: [Staff]
      summary: Opsi customer untuk select (admin, staff)
      description: Mengambil semua customer aktif tanpa pagination, search, atau limit untuk kebutuhan select transaksi.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
      responses:
        "200":
          description: Opsi customer berhasil diambil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerOptionListSuccessResponse"
              examples:
                success:
                  value:
                    data:
                      - id: ccustomer00000000000000001
                        name: Customer Satu
                        email: customer@example.com
                        account_number: "1000000001"
                    status: true
                    message: Opsi customer berhasil diambil
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/staff/customers/{customer_id}:
    get:
      tags: [Staff]
      summary: Detail customer (admin, staff)
      description: Mengambil detail core customer, rekening, status, dan mutasi terbaru.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/CustomerIdPath"
      responses:
        "200":
          $ref: "#/components/responses/CustomerDetailSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    put:
      tags: [Staff]
      summary: Update customer (admin, staff)
      description: Memperbarui profil, status, dan opsional password customer.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/CustomerIdPath"
      requestBody:
        $ref: "#/components/requestBodies/CustomerBody"
      responses:
        "200":
          $ref: "#/components/responses/CustomerSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [Staff]
      summary: Hapus customer (admin, staff)
      description: Soft delete customer dan menandai status menjadi inactive.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/CustomerIdPath"
      responses:
        "200":
          $ref: "#/components/responses/EmptySuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/staff/mutations:
    get:
      tags: [Staff]
      summary: List semua mutasi customer (admin, staff)
      description: |
        Mengambil mutasi semua customer dan menyertakan aktor yang melakukan transaksi.

        role = `admin`, `staff`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PerPageQuery"
        - $ref: "#/components/parameters/TransactionTypeQuery"
        - $ref: "#/components/parameters/TransactionDirectionQuery"
        - $ref: "#/components/parameters/CustomerIdQuery"
        - $ref: "#/components/parameters/ActorRoleQuery"
        - $ref: "#/components/parameters/DateFromQuery"
        - $ref: "#/components/parameters/DateToQuery"
      responses:
        "200":
          $ref: "#/components/responses/AllMutationListSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/staff/mutations/{id}:
    get:
      tags: [Staff]
      summary: Detail mutasi (admin, staff)
      description: Mengambil detail satu mutasi berdasarkan ID transaksi.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/MutationIdPath"
      responses:
        "200":
          $ref: "#/components/responses/AllMutationSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/staff/customers/{customer_id}/deposit:
    post:
      tags: [Staff]
      summary: Setor tunai untuk customer (admin, staff)
      description: |
        Staff atau admin melakukan setor tunai dengan memilih customer.

        role = `admin`, `staff`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/CustomerIdPath"
      requestBody:
        $ref: "#/components/requestBodies/AmountBody"
      responses:
        "201":
          $ref: "#/components/responses/StaffDepositSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/staff/customers/{customer_id}/withdraw:
    post:
      tags: [Staff]
      summary: Tarik tunai untuk customer (admin, staff)
      description: |
        Staff atau admin melakukan tarik tunai dengan memilih customer.

        role = `admin`, `staff`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/CustomerIdPath"
      requestBody:
        $ref: "#/components/requestBodies/AmountBody"
      responses:
        "201":
          $ref: "#/components/responses/StaffWithdrawSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/staff/customers/{customer_id}/transfer:
    post:
      tags: [Staff]
      summary: Transfer untuk customer (admin, staff)
      description: |
        Staff atau admin melakukan transfer dari customer pilihan ke customer tujuan.

        role = `admin`, `staff`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/CustomerIdPath"
      requestBody:
        $ref: "#/components/requestBodies/CustomerTransferBody"
      responses:
        "201":
          $ref: "#/components/responses/StaffTransferSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Gunakan header `Authorization: Bearer <access_token>`."
  parameters:
    AuthorizationHeader:
      name: Authorization
      in: header
      required: true
      description: Token bearer hasil login.
      schema:
        type: string
        example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example.signature
    SearchQuery:
      name: q
      in: query
      required: false
      description: Kata kunci pencarian list.
      schema:
        type: string
        example: customer
    PageQuery:
      name: page
      in: query
      required: false
      description: Nomor halaman. Default `1`.
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    PerPageQuery:
      name: per_page
      in: query
      required: false
      description: Jumlah data per halaman. Default `10`, maksimal `100`.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
        example: 10
    CustomerIdPath:
      name: customer_id
      in: path
      required: true
      description: ID user customer tujuan operasi.
      schema:
        type: string
        example: ccustomer00000000000000001
...
    MutationIdPath:
      name: id
      in: path
      required: true
      description: ID mutasi/transaksi.
      schema:
        type: string
        example: ctx000000000000000000001
    CustomerStatusQuery:
      name: status
      in: query
      required: false
      description: Filter status customer.
      schema:
        $ref: "#/components/schemas/CustomerStatus"
    TransactionTypeQuery:
      name: type
      in: query
      required: false
      description: Filter jenis transaksi.
      schema:
        $ref: "#/components/schemas/TransactionType"
    TransactionDirectionQuery:
      name: direction
      in: query
      required: false
      description: Filter arah transaksi.
      schema:
        $ref: "#/components/schemas/TransactionDirection"
    CustomerIdQuery:
      name: customer_id
      in: query
      required: false
      description: Filter mutasi berdasarkan customer.
      schema:
        type: string
        example: ccustomer00000000000000001
    ActorRoleQuery:
      name: actor_role
      in: query
      required: false
      description: Filter role aktor transaksi.
      schema:
        $ref: "#/components/schemas/Role"
    DateFromQuery:
      name: date_from
      in: query
      required: false
      description: Tanggal awal filter mutasi (inklusif).
      schema:
        type: string
        format: date
        example: "2026-07-01"
    DateToQuery:
      name: date_to
      in: query
      required: false
      description: Tanggal akhir filter mutasi (inklusif).
      schema:
        type: string
        format: date
        example: "2026-07-31"
  requestBodies:
    AmountBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AmountRequest"
          examples:
            default:
              value:
                amount: "100000.00"
                note: Transaksi teller
...
    CustomerBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CustomerUpsertRequest"
          examples:
            create:
              value:
                name: Customer Baru
                email: customerbaru@example.com
                password: password123
                status: active
  responses:
    ErrorResponse:
      description: Contoh response error standar. Status code aktual dapat berupa 400, 401, 403, 404, 409, 502, atau 500 sesuai kasus.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            error:
              value:
                status: false
                message: Request tidak valid
    CustomerListSuccessResponse:
      description: Customer berhasil diambil
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CustomerListSuccessResponse"
          examples:
            success:
              value:
                data:
                  items:
                    - id: ccustomer00000000000000001
                      name: Customer Satu
                      email: customer@example.com
                      account_number: "1000000001"
                      balance: 98750000
                      status: active
                      created_at: "2026-07-23 15:30:00"
                      updated_at: "2026-07-23 15:30:00"
                  pagination:
                    current_page: 1
                    per_page: 10
                    from: 1
                    total: 1
                    last_page: 1
                status: true
                message: Customer berhasil diambil
    AllMutationListSuccessResponse:
      description: Mutasi semua customer berhasil diambil
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AllMutationListSuccessResponse"
          examples:
            success:
              value:
                data:
                  items:
                    - id: ctx000000000000000000001
                      customer_id: ccustomer00000000000000001
                      customer_name: Customer Satu
                      actor_name: Staff Teller
                      actor_role: staff
                      type: deposit
                      direction: in
                      amount: 10000000
                      balance_before: 90000000
                      balance_after: 100000000
                      reference_id: null
                      note: Setor tunai teller
                      created_at: "2026-07-23 15:30:00"
                  pagination:
                    current_page: 1
                    per_page: 10
                    from: 1
                    total: 1
                    last_page: 1
                status: true
                message: Mutasi semua customer berhasil diambil
...
    CustomerSuccessResponse:
      description: Customer berhasil diproses
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CustomerSuccessResponse"
    CustomerDetailSuccessResponse:
      description: Detail customer berhasil diambil
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CustomerDetailSuccessResponse"
...
    AllMutationSuccessResponse:
      description: Detail mutasi semua customer berhasil diambil
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AllMutationSuccessResponse"
    DashboardSuccessResponse:
      description: Ringkasan dashboard berhasil diambil
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DashboardSuccessResponse"
  schemas:
...
    CustomerOptionListSuccessResponse:
      type: object
      required: [data, status, message]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CustomerOptionItem"
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Opsi customer berhasil diambil
...
    DashboardSuccessResponse:
      type: object
      required: [data, status, message]
      properties:
        data:
          $ref: "#/components/schemas/DashboardData"
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Ringkasan dashboard berhasil diambil
...